added map differences depending on location
