+++ controllers/api/index.js
@@ -3,9 +3,11 @@ const router = require('express').Router();
 const userRoutes = require('./user-routes.js');
 const postRoutes = require('./post-routes');
 const commentRoutes = require('./comment-routes');
+const locationRoutes = require('./location-routes');
 
 router.use('/users', userRoutes);
 router.use('/posts', postRoutes);
 router.use('/comments', commentRoutes);
+router.use('/locations', locationRoutes);
 
 module.exports = router;
