+++ public/javascript/logout.js
@@ -1,6 +1,3 @@
-
-
-
 async function logout() {
 	const response = await fetch('/api/users/logout', {
 		method: 'post',
