+++ views/dashboard.handlebars
@@ -1,1 +1,3 @@
 <h1>Dashboard</h1>
+
+{{>map}}
