+++ views/homepage.handlebars
@@ -10,7 +10,7 @@
                 <p class="post-byline"><span class="post-user">{{user.username}}</span> | {{format_date created_at}} ({{timeSince created_at}})</p>
 
             </div>
-            <h4>comments</h4>
+            <h4>{{commentCount post.comments}}</h4>
             {{#each comments as |comment|}}
                 <p class="comment-text">{{comment_text}}</p>
                 <p class="comment-byline"><span class="comment-user">{{user.username}}</span> | {{format_date created_at}} ({{timeSince created_at}})</p>
